+++ src/containerapp/azext_containerapp/_help.py
@@ -109,6 +109,7 @@ helps['containerapp service'] = """
     type: group
     short-summary: Commands to manage services available within the environment.
 <<<<<<< HEAD
+<<<<<<< HEAD
     deprecate_info: This command group is deprecated. Use 'az containerapp add-on' instead.
 """
 
@@ -117,12 +118,19 @@ helps['containerapp add-on'] = """
     short-summary: Commands to manage add-ons available within the environment.
 =======
     deprecate_info: This command group is deprecated. Use 'az containerapp addon' instead.
+=======
+    deprecate_info: This command group is deprecated. Use 'az containerapp add-on' instead.
+>>>>>>> a9dc64ae6 (add-on)
 """
 
-helps['containerapp addon'] = """
+helps['containerapp add-on'] = """
     type: group
+<<<<<<< HEAD
     short-summary: Commands to manage addons available within the environment.
 >>>>>>> 9d479eb49 (add)
+=======
+    short-summary: Commands to manage add-ons available within the environment.
+>>>>>>> a9dc64ae6 (add-on)
 """
 
 helps['containerapp service list'] = """
@@ -130,6 +138,7 @@ helps['containerapp service list'] = """
     short-summary: List all services within the environment.
 """
 
+<<<<<<< HEAD
 <<<<<<< HEAD
 helps['containerapp add-on list'] = """
     type: command
@@ -139,6 +148,11 @@ helps['containerapp addon list'] = """
     type: command
     short-summary: List all addons within the environment.
 >>>>>>> 9d479eb49 (add)
+=======
+helps['containerapp add-on list'] = """
+    type: command
+    short-summary: List all add-ons within the environment.
+>>>>>>> a9dc64ae6 (add-on)
 """
 
 helps['containerapp service redis'] = """
@@ -217,6 +231,7 @@ helps['containerapp service qdrant delete'] = """
 """
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 helps['containerapp add-on redis'] = """
     type: group
     short-summary: Commands to manage the redis add-on for the Container Apps environment.
@@ -292,79 +307,86 @@ helps['containerapp add-on qdrant delete'] = """
     short-summary: Command to delete the qdrant add-on.
 =======
 helps['containerapp addon redis'] = """
+=======
+helps['containerapp add-on redis'] = """
+>>>>>>> a9dc64ae6 (add-on)
     type: group
-    short-summary: Commands to manage the redis addon for the Container Apps environment.
+    short-summary: Commands to manage the redis add-on for the Container Apps environment.
 """
 
-helps['containerapp addon postgres'] = """
+helps['containerapp add-on postgres'] = """
     type: group
-    short-summary: Commands to manage the postgres addon for the Container Apps environment.
+    short-summary: Commands to manage the postgres add-on for the Container Apps environment.
 """
 
-helps['containerapp addon kafka'] = """
+helps['containerapp add-on kafka'] = """
     type: group
-    short-summary: Commands to manage the kafka addon for the Container Apps environment.
+    short-summary: Commands to manage the kafka add-on for the Container Apps environment.
 """
 
-helps['containerapp addon mariadb'] = """
+helps['containerapp add-on mariadb'] = """
     type: group
-    short-summary: Commands to manage the mariadb addon for the Container Apps environment.
+    short-summary: Commands to manage the mariadb add-on for the Container Apps environment.
 """
 
-helps['containerapp addon qdrant'] = """
+helps['containerapp add-on qdrant'] = """
     type: group
-    short-summary: Commands to manage the qdrant addon for the Container Apps environment.
+    short-summary: Commands to manage the qdrant add-on for the Container Apps environment.
 """
 
-helps['containerapp addon redis create'] = """
+helps['containerapp add-on redis create'] = """
     type: command
-    short-summary: Command to create the redis addon.
+    short-summary: Command to create the redis add-on.
 """
 
-helps['containerapp addon postgres create'] = """
+helps['containerapp add-on postgres create'] = """
     type: command
-    short-summary: Command to create the postgres addon.
+    short-summary: Command to create the postgres add-on.
 """
 
-helps['containerapp addon kafka create'] = """
+helps['containerapp add-on kafka create'] = """
     type: command
-    short-summary: Command to create the kafka addon.
+    short-summary: Command to create the kafka add-on.
 """
 
-helps['containerapp addon mariadb create'] = """
+helps['containerapp add-on mariadb create'] = """
     type: command
-    short-summary: Command to create the mariadb addon.
+    short-summary: Command to create the mariadb add-on.
 """
 
-helps['containerapp addon qdrant create'] = """
+helps['containerapp add-on qdrant create'] = """
     type: command
-    short-summary: Command to create the qdrant addon.
+    short-summary: Command to create the qdrant add-on.
 """
 
-helps['containerapp addon redis delete'] = """
+helps['containerapp add-on redis delete'] = """
     type: command
-    short-summary: Command to delete the redis addon.
+    short-summary: Command to delete the redis add-on.
 """
 
-helps['containerapp addon postgres delete'] = """
+helps['containerapp add-on postgres delete'] = """
     type: command
-    short-summary: Command to delete the postgres addon.
+    short-summary: Command to delete the postgres add-on.
 """
 
-helps['containerapp addon kafka delete'] = """
+helps['containerapp add-on kafka delete'] = """
     type: command
-    short-summary: Command to delete the kafka addon.
+    short-summary: Command to delete the kafka add-on.
 """
 
-helps['containerapp addon mariadb delete'] = """
+helps['containerapp add-on mariadb delete'] = """
     type: command
-    short-summary: Command to delete the mariadb addon.
+    short-summary: Command to delete the mariadb add-on.
 """
 
-helps['containerapp addon qdrant delete'] = """
+helps['containerapp add-on qdrant delete'] = """
     type: command
+<<<<<<< HEAD
     short-summary: Command to delete the qdrant addon.
 >>>>>>> 9d479eb49 (add)
+=======
+    short-summary: Command to delete the qdrant add-on.
+>>>>>>> a9dc64ae6 (add-on)
 """
 
 helps['containerapp env update'] = """