+++ src/containerapp/azext_containerapp/commands.py
@@ -52,6 +52,7 @@ def load_command_table(self, _):
 
 <<<<<<< HEAD
 <<<<<<< HEAD
+<<<<<<< HEAD
     with self.command_group('containerapp service', deprecate_info=self.deprecate(redirect='containerapp add-on', hide=True), is_preview=True) as g:
         g.custom_command('list', 'list_all_services')
 
@@ -100,49 +101,56 @@ def load_command_table(self, _):
 =======
     with self.command_group('containerapp service', deprecate_info=self.deprecate(redirect='containerapp addon', hide=True), is_preview=True) as g:
 >>>>>>> 6fbd6b754 (recordings live)
+=======
+    with self.command_group('containerapp service', deprecate_info=self.deprecate(redirect='containerapp add-on', hide=True), is_preview=True) as g:
+>>>>>>> a9dc64ae6 (add-on)
         g.custom_command('list', 'list_all_services')
 
-    with self.command_group('containerapp addon', is_preview=True) as g:
+    with self.command_group('containerapp add-on', is_preview=True) as g:
         g.custom_command('list', 'list_all_services')
 
-    with self.command_group('containerapp service redis', deprecate_info=self.deprecate(redirect='containerapp addon redis', hide=True)) as g:
+    with self.command_group('containerapp service redis', deprecate_info=self.deprecate(redirect='containerapp add-on redis', hide=True)) as g:
         g.custom_command('create', 'create_redis_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_redis_service', confirmation=True, supports_no_wait=True)
 
-    with self.command_group('containerapp addon redis') as g:
+    with self.command_group('containerapp add-on redis') as g:
         g.custom_command('create', 'create_redis_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_redis_service', confirmation=True, supports_no_wait=True)
 
-    with self.command_group('containerapp service postgres', deprecate_info=self.deprecate(redirect='containerapp addon postgres', hide=True)) as g:
+    with self.command_group('containerapp service postgres', deprecate_info=self.deprecate(redirect='containerapp add-on postgres', hide=True)) as g:
         g.custom_command('create', 'create_postgres_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_postgres_service', confirmation=True, supports_no_wait=True)
 
-    with self.command_group('containerapp addon postgres') as g:
+    with self.command_group('containerapp add-on postgres') as g:
         g.custom_command('create', 'create_postgres_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_postgres_service', confirmation=True, supports_no_wait=True)
 
-    with self.command_group('containerapp service kafka', deprecate_info=self.deprecate(redirect='containerapp addon kafka', hide=True)) as g:
+    with self.command_group('containerapp service kafka', deprecate_info=self.deprecate(redirect='containerapp add-on kafka', hide=True)) as g:
         g.custom_command('create', 'create_kafka_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_kafka_service', confirmation=True, supports_no_wait=True)
 
-    with self.command_group('containerapp addon kafka') as g:
+    with self.command_group('containerapp add-on kafka') as g:
         g.custom_command('create', 'create_kafka_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_kafka_service', confirmation=True, supports_no_wait=True)
 
-    with self.command_group('containerapp service mariadb', deprecate_info=self.deprecate(redirect='containerapp addon mariadb', hide=True)) as g:
+    with self.command_group('containerapp service mariadb', deprecate_info=self.deprecate(redirect='containerapp add-on mariadb', hide=True)) as g:
         g.custom_command('create', 'create_mariadb_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_mariadb_service', confirmation=True, supports_no_wait=True)
 
-    with self.command_group('containerapp addon mariadb') as g:
+    with self.command_group('containerapp add-on mariadb') as g:
         g.custom_command('create', 'create_mariadb_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_mariadb_service', confirmation=True, supports_no_wait=True)
 
-    with self.command_group('containerapp service qdrant', deprecate_info=self.deprecate(redirect='containerapp addon qdrant', hide=True)) as g:
+    with self.command_group('containerapp service qdrant', deprecate_info=self.deprecate(redirect='containerapp add-on qdrant', hide=True)) as g:
         g.custom_command('create', 'create_qdrant_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_qdrant_service', confirmation=True, supports_no_wait=True)
 
+<<<<<<< HEAD
     with self.command_group('containerapp addon qdrant') as g:
 >>>>>>> 9d479eb49 (add)
+=======
+    with self.command_group('containerapp add-on qdrant') as g:
+>>>>>>> a9dc64ae6 (add-on)
         g.custom_command('create', 'create_qdrant_service', supports_no_wait=True)
         g.custom_command('delete', 'delete_qdrant_service', confirmation=True, supports_no_wait=True)